pyproject.toml
[project]
name = "rollout-runner"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "sqlmodel>=0.0.21",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
